requirements.txt
# NautilusTrader Core (1.221.0 is latest for Python 3.11)
nautilus_trader[docker]==1.221.0

# Interactive Brokers API - installed from official source in Dockerfile

# RL Training
stable-baselines3>=2.3.0
gymnasium>=0.29.0
finrl>=0.3.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
polars>=0.20.0

# Data Providers (FREE)
yfinance>=0.2.36
requests>=2.31.0
aiohttp>=3.9.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# TimescaleDB
timescale-vector>=0.0.1

# ML & Tracking
mlflow>=2.10.0
torch>=2.0.0

# Telegram
python-telegram-bot>=21.0

# Monitoring
prometheus-client>=0.19.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
httpx>=0.26.0
aiofiles>=23.0.0
structlog>=24.0.0
tenacity>=8.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
